rename components v2
